requirements.txt
asn1crypto==0.24.0
cffi==1.12.2
Click==7.0
cryptography==2.6.1
entrypoints==0.3
flake8==3.7.7
Flask==1.0.2
Flask-Fixtures==0.3.8
flask-marshmallow==0.10.0
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==2.19.2
marshmallow-sqlalchemy==0.16.1
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
python-dateutil==2.8.0
python-dotenv==0.10.1
six==1.12.0
SQLAlchemy==1.3.2
Werkzeug==0.15.2
